requirements.txt
python==3.11.3
django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
